requirements.txt
django
django-extra-views
Pillow
django-bootstrap-form
django-mathfilters
